requirements.txt
paho-mqtt
Pillow
python-dotenv
requests
wg-utilities
